+++ plugins/decl_netcdf/tests/decl_netcdf_tests.cxx
@@ -1,6 +1,6 @@
 /*******************************************************************************
  * Copyright (C) 2020-2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
- * Copyright (C) 2024-2025 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2024-2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -847,6 +847,178 @@ TEST(decl_netcdf_test, size_of)
 		// printf("%d ?= %d\n", int_array[i], i);
 		ASSERT_EQ(int_array[i], i);
 	}
+
+	PDI_finalize();
+}
+
+/*
+ * Name:                decl_netcdf_test.defalte
+ *
+ * Description:         Tests simple write and read of compressed variables
+*/
+TEST(decl_netcdf_test, deflate)
+{
+	constexpr char CONFIG_YAML[] = R"(
+    logging: trace
+    metadata:
+      pb_size: int
+      input: int
+      chunk: int
+    data:
+      int_scalar: int
+      int_array: {type: array, subtype: int, size: $pb_size}
+      int_matrix:
+        type: array
+        subtype: int
+        size: ['$pb_size', '$pb_size']
+    plugins:
+      decl_netcdf:
+      - file: 'test_deflate_0.nc'
+        variables:
+          int_scalar: int
+          int_array:
+            type: array
+            subtype: int
+            size: $pb_size
+            dimensions: ['time']
+          int_matrix:
+            type: array
+            subtype: int
+            size: ['$pb_size', '$pb_size']
+            dimensions: ['col', 'row']
+        when: '${input}=0'
+        write: [int_scalar, int_array, int_matrix]
+      - file: 'test_deflate_6.nc'
+        variables:
+          int_scalar: int
+          int_array:
+            type: array
+            subtype: int
+            size: $pb_size
+            dimensions: ['time']
+            deflate: 6
+          int_matrix:
+            type: array
+            subtype: int
+            size: ['$pb_size', '$pb_size']
+            dimensions: ['col', 'row']
+            deflate: 6
+        when: '${input}=0'
+        write: [int_scalar, int_array, int_matrix]
+      - file: 'test_deflate_9.nc'
+        deflate: 9
+        variables:
+          int_scalar: int
+          int_array:
+            type: array
+            subtype: int
+            size: $pb_size
+            dimensions: ['time']
+            chunking: $chunk
+          int_matrix:
+            type: array
+            subtype: int
+            size: ['$pb_size', '$pb_size']
+            dimensions: ['col', 'row']
+            chunking: ['$chunk', '$chunk']
+        when: '${input}=0'
+        write: [int_scalar, int_array, int_matrix]
+      - file: 'test_deflate_mix.nc'
+        deflate: 6
+        variables:
+          int_scalar: int
+          int_array:
+            type: array
+            subtype: int
+            size: $pb_size
+            dimensions: ['time']
+            deflate: 9
+            chunking: $chunk
+          int_matrix:
+            type: array
+            subtype: int
+            size: ['$pb_size', '$pb_size']
+            dimensions: ['col', 'row']
+            chunking: ['$chunk', '$chunk']
+        when: '${input}=0'
+        write: [int_scalar, int_array, int_matrix]
+      - file: 'test_deflate_6.nc'
+        when: '${input}=1'
+        read: [int_scalar, int_array, int_matrix]
+    )";
+
+	PDI_init(PC_parse_string(CONFIG_YAML));
+
+	// init data
+	int input = 0;
+	int int_scalar = 42;
+	int N = 1000;
+	int chunk = 1000;
+	int* int_array = new int[N];
+	int* int_matrix = new int[N * N];
+
+	// write data
+	for (int i = 0; i < N; i++) {
+		int_array[i] = i;
+	}
+
+	for (int i = 0; i < N * N; i++) {
+		int_matrix[i] = i;
+	}
+
+	PDI_expose("input", &input, PDI_OUT);
+	PDI_expose("pb_size", &N, PDI_OUT);
+	PDI_expose("chunk", &chunk, PDI_OUT);
+
+	PDI_expose("int_scalar", &int_scalar, PDI_OUT);
+	PDI_expose("int_array", int_array, PDI_OUT);
+	PDI_expose("int_matrix", int_matrix, PDI_OUT);
+
+	// check the deflate level of output files
+	int result;
+	result = std::system("which ncdump > /dev/null 2>&1");
+	// check the deflate level only if ncdump is available
+	if (result == 0) {
+		result = std::system("ncdump -hs test_deflate_6.nc | grep -q '_DeflateLevel = 6'");
+		EXPECT_EQ(result, 0) << "Deflate level for test_deflate_6.nc is not 6";
+		result = std::system("ncdump -hs test_deflate_9.nc | grep -q '_DeflateLevel = 9'");
+		EXPECT_EQ(result, 0) << "Deflate level for test_deflate_9.nc is not 9";
+		result = std::system("ncdump -hs test_deflate_mix.nc | grep -q 'int_array:_DeflateLevel = 9'");
+		EXPECT_EQ(result, 0) << "Deflate level of int_array in test_deflate_mix.nc is not 9";
+		result = std::system("ncdump -hs test_deflate_mix.nc | grep -q 'int_matrix:_DeflateLevel = 6'");
+		EXPECT_EQ(result, 0) << "Deflate level of int_matrix in test_deflate_mix.nc is not 6";
+	}
+
+
+	// read data
+	input = 1;
+	int_scalar = 0;
+	for (int i = 0; i < N; i++) {
+		int_array[i] = 0;
+	}
+	for (int i = 0; i < N * N; i++) {
+		int_matrix[i] = 0;
+	}
+
+	PDI_expose("input", &input, PDI_OUT);
+
+	PDI_expose("int_scalar", &int_scalar, PDI_IN);
+	PDI_expose("int_array", int_array, PDI_IN);
+	PDI_expose("int_matrix", int_matrix, PDI_IN);
+
+	// verify
+	ASSERT_EQ(int_scalar, 42);
+
+	for (int i = 0; i < N; i++) {
+		ASSERT_EQ(int_array[i], i);
+	}
+
+	for (int i = 0; i < N * N; i++) {
+		ASSERT_EQ(int_matrix[i], i);
+	}
+
+	delete[] int_array;
+	delete[] int_matrix;
 
 	PDI_finalize();
 }